form

+++ src/mixins/consoleCudr.js
@@ -23,6 +23,13 @@ export default {
     handleRowEdit(scope) {
       this.$refs.DialogRow.handleOpen(deepClone(scope.row), 'EDIT')
     },
+
+    handleRowDetail(scope) {
+      this.$router.push({
+        path: `${scope.row.id}`,
+        append: true
+      })
+    },
     // 跳转至页面编辑
     handleRowEditpage(scope) {
       this.$router.push({
@@ -32,7 +39,7 @@ export default {
     },
 
     handleUpdate(form) {
-      return this.updateApi(`/${this.apiName}`, form)
+      return this.updateApi(`/${this.API_NAME}`, form)
     },
 
     handleMultipleAction(e) {
@@ -56,13 +63,21 @@ export default {
         this[`handle${e}`]()
       } else {
         const { mode = 'Row', command, scope } = e
-        this[`handle${mode}${command}`](scope)
+        if (mode === 'Row') {
+          this[`handleRow${command}`](scope)
+        } else {
+          this.handleMultipleAction(e)
+        }
       }
     },
-
+    // 行删除
     handleRowDelete(scope) {
       const { row } = scope
-      this.handleDelete(row.id)
+      this.$confirm('确认操作?', '提示', {
+        type: 'warning'
+      }).then(() => {
+        this.handleDelete(row.id)
+      })
     },
 
     async handleDelete(id) {
@@ -70,20 +85,20 @@ export default {
         console.warn('no-id')
         return
       }
-      await this.Fetch.delete(`/${this.apiName}/${id}`)
+      await this.Fetch.delete(`/${this.API_NAME}/${id}`)
       this.actionSuccess()
       this.init()
     },
 
     async handleRowSubmit(form) {
-      try {
-        await this.handleUpdate(form)
-      } catch (e) {
-        return
-      }
-      this.$refs.DialogRow.handleClose()
-      this.actionSuccess()
-      this.init()
+      // try {
+      //   await this.handleUpdate(form)
+      // } catch (e) {
+      //   return
+      // }
+      // this.$refs.DialogRow.handleClose()
+      // this.actionSuccess()
+      // this.init()
     },
 
     handleRowClone(scope) {
